discord-bot/requirements.txt
discord==2.3.2
discord.py==2.6.3
google-genai==1.19.0
python-dotenv==1.0.1
supabase==2.15.2
sentistrength==0.0.9
google==3.0.0